+++ spec/git.spec.js
@@ -4,11 +4,11 @@ var git = require( "../src/git.js" );
 var repoPath = path.resolve( "./spec/fauxgitaboudit" );
 
 describe( "Git", function() {
-	describe( "when getting basic information", function() {
+	describe( "when getting basic information with default tag", function() {
 		var repoInfo;
 
 		before( function( done ) {
-			git.repo( repoPath )
+			git.repo( repoPath, "o_r_b_v_c_s" )
 				.then( function( info ) {
 					repoInfo = info;
 					done();
@@ -22,7 +22,43 @@ describe( "Git", function() {
 			repoInfo.path.should.equal( repoPath );
 			repoInfo.build.should.equal( 5 );
 			repoInfo.commit.length.should.equal( 40 );
+			repoInfo.major.should.equal( "0" );
+			repoInfo.minor.should.equal( "1" );
+			repoInfo.patch.should.equal( "1" );
+			repoInfo.major_version.should.equal( "0" );
+			repoInfo.minor_version.should.equal( "0.1" );
 			repoInfo.tag.should.equal( "anonymous_fauxgitaboudit_master_" + repoInfo.version + "_5_" + repoInfo.commit.slice( 0, 8 ) );
+		} );
+	} );
+
+	describe( "when getting basic information with multiple tags", function() {
+		var repoInfo;
+
+		before( function( done ) {
+			git.repo( repoPath, "v_c_s,miv,ma" )
+				.then( function( info ) {
+					repoInfo = info;
+					done();
+				} );
+		} );
+
+		it( "should retrieve necessary repository data from environment", function() {
+			repoInfo.owner.should.equal( "anonymous" );
+			repoInfo.repository.should.equal( "fauxgitaboudit" );
+			repoInfo.branch.should.equal( "master" );
+			repoInfo.path.should.equal( repoPath );
+			repoInfo.build.should.equal( 5 );
+			repoInfo.commit.length.should.equal( 40 );
+			repoInfo.major.should.equal( "0" );
+			repoInfo.minor.should.equal( "1" );
+			repoInfo.patch.should.equal( "1" );
+			repoInfo.major_version.should.equal( "0" );
+			repoInfo.minor_version.should.equal( "0.1" );
+			repoInfo.tag.should.eql( [
+				repoInfo.version + "_5_" + repoInfo.commit.slice( 0, 8 ),
+				"0.1",
+				"0"
+			] );
 		} );
 	} );
 
@@ -32,7 +68,7 @@ describe( "Git", function() {
 		before( function() {
 			process.env.DRONE = true;
 			process.env.DRONE_BRANCH = "testing";
-			return git.repo( repoPath )
+			return git.repo( repoPath, "o_r_b_v_c_s" )
 				.then( function( info ) {
 					repoInfo = info;
 				} );