+++ src/command.js
@@ -1,11 +1,10 @@
-var _ = require( "lodash" );
 var exec = require( "child_process" ).exec;
 var when = require( "when" );
 
 function executeCommand( line, path ) {
 	path = path || __dirname;
 	return when.promise( function( resolve, reject ) {
-		var command = _.isArray( line ) ? line.join( " " ) : line;
+		var command = Array.isArray( line ) ? line.join( " " ) : line;
 		exec( command,
 			{ cwd: path },
 			function( err, stdout /*, stderr */ ) {