+++ src/index.js
@@ -2,6 +2,7 @@ var fs = require( "fs" );
 var path = require( "path" );
 var root = process.argv.length > 2 ? path.resolve( process.argv[ 2 ] ) : path.resolve( process.cwd() );
 var format = require( "util" ).format;
+var argv = require( "yargs" ).default( { tag: "o_r_b_v_c_s" } ).argv;
 var git = require( "./git" );
 
 console.log( format( "Getting build information for repository at '%s'", root ) );
@@ -22,5 +23,5 @@ function onError( err ) {
 	console.error( "Could not get build information due to", message );
 }
 
-git.repo( root )
+git.repo( root, argv.tag )
 	.then( onInfo, onError );