+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.x
+
+## 0.2.0
+
+* Adds support for custom tag format
+* Adds support for multiple tags
+* Removes lodash dependency
+
 ## 0.1.x
 
 ### 0.1.5